refactor(muya): centralize code fence language derivation (review cleanup)
Applies /simplify review feedback to the info-string refactor:

- Derive the highlight/tokenize language once in `CodeBlockContent._lang`
  (returns the info string's first word) instead of at each call site. This
  fixes two consumers the scattered approach missed — the tabHandler
  markup test and the Firefox-compat backspaceHandler both read the raw
  field and silently mis-derived the language for a multi-word info string.
- Move `firstWordOfInfo` from block/commonMark/codeBlock into utils, so the
  state layer no longer reaches into the block tree for it (removes a
  reverse state -> block runtime import).
- Simplify `lang: isFenced ? info : lang` to `lang: info` in
  `_buildCodeState` — indented blocks have no info string, so both arms are
  equal.
